+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-## v0.10.0 / 2017-02-xx
+## v0.10.0 / 2017-xx-xx
 - Remove console.log debug code
-- Implemented array aggregation operators       
+- Implemented array aggregation operators
   (`$arrayElemAt`,`$concatArrays`,`$filter`,`$indexOfArray`,`$isArray`,`$range`,`$reverseArray`,`$reduce`,`$slice`,`$zip`)
 - Implemented string aggregation operators (`$indexOfBytes`,`$split`)
+- Implemented arithmetic aggregation operators (`$ceil`,`$exp`,`$floor`,`$ln`,`$log`,`$log10`,`$pow`,`$sqrt`,`$trunc`)
 - Added .editorconfig
 - Pass utility functions to custom operator implementation
 - Rename function to retrieve collection id to `idKey` in custom operators